ALI files: consolidate the file kinds of main sources for C
Main sources for C are looked up from "D line" entries, which load files
as stubs. Make sure we consolidate them to source files when we decide
to treat them as source files.

This fixes a crash in the homonym-units-of-interest testcase that the
eng/cov/gnatcoverage!8 MR creates.

TN: V728-002