sc_obligations.adb: fix rules to discard CUs from checkpoints
CU have two file attributes: Origin (the file from which the SCOs are
computed/loaded: LI files for binary traces, and source files themselves
for source traces).

While so far, a CU loaded from a checkpoint was discarded when Origin
was an ignored source file, we used to try to load it anyway even if the
Main_Source was ignored, which is invalid and resulted in a crash. This
commit fixes this bug: it discards the CU if Origin *or* Main_Source are
ignored.

For eng/cov/gnatcoverage#17