files_table.adb: do not call Simple_Name on Clang predefined file names
Following recent changes for for eng/cov/gnatcoverage#17, the
tests-instr-cov-cmdline-self-ref-macro testcase started failing on
Windows: the problem is that we inserted a call to
Ada.Directories.Simple_Name on a file name that may be artificial:
"<command line>" is not real file name, and is used to designate source
that appears in the command line (not from a source file). This name is
invalid on Windows, and thus Simple_Name raises a Name_Error exception
for it on Windows.