testsuite: evolve warnings_as_errors into tolerate_messages
Replace the "warnings_as_errors" bool argument to xcov_instrument
by a "tolerate_messages" regular expression for which matching warnings
or errors shouldn't result in a test failure.

This allows finer grain expression of what tests expect and
addresses the slightly misleading aspect of the previous parameter
name, which was also preventing failures on _errors_ when
set to False.

Tests in this repo are adjusted to new API and similar adjustments
are to be provided for tests hosted in separate repositories.

A secondary motivation for this change is to facilitate supporting
tests doing instrumentation before builds without complaining about
missing object dirs, in qualification contexts possibly using old
versions of gnatcov.

This part will come as a separate change.

Part of W215-023 (gnatcov 22.2 qualification)