Instrument: fix a crash when Libadalang cannot find the parent decl
This commit fixes a crash in the Ada instrumenter. The crash occurred
when instrumenting an incomplete code closure: Libadalang can return a
null node when it cannot find the "parent" declaration of the current
unit (package Pkg for package Pkg.Child, for instance).

In the new testcase, whether or not "gnatcov instrument" should skip the
instrumentation of pkg-child.adb (just like gprbuild skips it), we
should have defensive code to avoid the crash, and emit a warning
instead.

TN: W213-002