Instrument: remove occurrences of Inst_Context in unit instr process
This is preliminary work for the parallelized instrumentation.

We want to isolate the most we can the instrumentation of a single
source. The Inst_Context is shared between all of the source
instrumentations, and holds global information which holds us against
being able to instrument a single compilation unit at a time.

This commit removes the uses of the Inst_Context in the unit
instrumentation process (in Instrument.Ada_Unit and Instrument.C). In
practice, this means no behavior change.

Note that this Inst_Context was used in the Ada instrumentation process
to actually check that the parent unit of a separate unit is
instrumented. As assertions are deactivated in production mode, we make
the controversial choice to drop it.

Note that we also use this as an opportunity to actually isolate the
libadalang dependency to Instrument.Ada_Unit, as the Inst_Context is no
longer the one holding the Libadalang.Analysis_Context, but the
Ada_Unit_Inst_Context is.