Replace Pre by Precondition in test on disabled assert pragma
For mcdc, the gnatcov 22 instrumenter mistakenly tries to instrument
a disabled library level "Pre" pragma, only to warn that it cannot find
a spot to declare the mcdc state.

This is not a real problem as the pragma was requested to be disabled
anyway, just produces a warning which triggers a test failure.

As the use of aspects instead of pragmas is generally encouraged, we
expect pragmas not to be used much if at all in the first place. This
change simply replaces a lone Pre pragma by a Precondition one, functionally
equivalent and better recognized by gnatcov 22.

Part of W215-023 (gnatcov 22.x qualification)