Testsuite: Add options to change dump trigger and channel
This change adds two options that respectively change the
default dump trigger and default dump channel. This is
useful to ensure for future qualifications that we are
testing gnatcov with the correct set of options.

Forcing the dump trigger or dump channel can mess with the
logic of some of the internal tests, by moving the buffer dump
point (and thus changing what is covered or not) or violating
some assumptions concerning the location of traces.

Part of W215-023 (gnatcov 22.x qualification)