Fix instrumentation of projects mixing C / C++
The list of coverage buffers is stored in a compilation unit unique to
the project, but for each language of the project. In practice, this
meant that we had a gnatcov_rts_c-buffers-lists-<prj_name>.c file (for
C), and a gnatcov_rts_c-buffers-lists-<prj_name>.cpp file (for C++),
both having the same object name which was rejected by the builder.

As the list of coverage buffers is the same, there is no need to
generate a declaration for each language. We now generate this list for
one of the language supported, and then reuse the exported symbol when
dumping coverage buffers.