Robustify Qualif/...550_Loop/src/lib_loop_statements wrt optimization
Reorder statements to prevent the occurrence of a
no-code line at the end of a loop sequence. The no-code
was legit but is annoying to test against. Reordering
the statements helps that, doesn't invalidate the test
and doesn't hide any defect.

Take the opportunity to add a few comments in a common
spec, describing the testcase(s) organisation.

Part of W215-023 (gnatcov 22.x qualification)