instrument.adb: always instrument sources in the same order
When instrumenting a unit, sort the source files to instrument before
iterating on them so that the order for instrumentation is
deterministic. This is useful both to avoid spurious variations in
tests, and to reduce the chance that we reproduce bugs in one
configuration, but not on another.

TN: W213-002