Instrument.C: output clang diagnostics
If the parsing failed because of an error in the user code, we should
emit the diagnostics reported by clang to let the user know that
coverage analysis may be incomplete.

As we preprocess code with another compiler configuration than the clang
one, we may have compiler-specific code, notably in system-headers, so
we also need to filter out a lot of messages to avoid issuing spurious
warning messages. The heuristic picked is to only keep diagnostics whose
presumed location file is a file of interest (belonging to the project).
This could be revisited if there are still too many false positives (as
we will have false positives raised on compiler-specific user code).