Change the report produced for --annotate=html[+]
There should be no reason to pick the HTML format over the dynamic HTML
format. This commit switches the report produced when --annotate=html[+]
is passed to the gnatcov coverage command to actually produce a dhtml
report.

We choose to keep for now the static HTML version, and produce it when
the (undocumented) shtml[+] option is passed to the annotate switch.