Fix regression with 32 bits configurations
To have proper report messages, we need to postprocess the source files
at `gnatcov coverage` time to match expected source locations computed
by `gnatcov instrument`. This postprocessing was done by the
Postprocess_Source function defined in the Instrument.C unit. This unit
is now stubbed for 32-bits configurations (which we do not need
instrumentation for), and the stubbed version of the procedure was not
doing anything.

In practice, this meant that retrieving code lines from this
non-existing postprocess file yielded empty strings, resulting in
missing information from the report when the code resulted from a macro
expansion.