instrument-ada_unit.adb: emit a warning when we cannot instrument main
As revealed by the bug report in W328-016, aborting the process of
inserting dump of coverage buffers silently is not a good idea: the
instrumentation seemingly worked ("gnatcov instrument" exits with code 0
and without warnings) but the build fails later due to linking issues.

To avoid confusion in the future, make gnatcov emit a warning in such
cases.