instrument-ada_unit.adb: rework syntax fields access in rewritten trees
Currently, the `Libadalang.Rewriting` API provides a single primitive to
access the syntax field of a rewritten node:

```ada
function Child
  (Handle : Node_Rewriting_Handle;
   Index  : Natural) return Node_Rewriting_Handle;
```

This API is both inconvenient (one has to compute ``Index`` for the
requested field, which depends on the actual node type for ``Handle``)
and error-prone (no guarantee when calling ``Child`` that the passed
index is consistent with the actual node type).

This refactoring creates a parallel set of primitives wrapping the
above. They accept a syntax field reference instead of the index, and
compute the index from the actual node type, which fixes both of the
issues described above.