instrument-ada_unit.adb: fix handling of mains that are specs
Assuming that `Main` is a generic procedure instantiation acting as a
main for the project to instrument:

```ada
procedure [Main] is new ...
```

In order to insert code to automatically dump coverage buffers
(`--dump-trigger=` anything but manual), we need a procedure body so
that we can insert dumps for coverage buffers. We can do this with a
wrapper procedure:

```ada
with Original_[Main];

procedure [Main] is
begin
   Original_[Main];
end [Main];
```

However we also need `Main` to still be a spec-only unit, so that the
GPR clause `for Main use ("[main].ads");` stays valid. To satisfy all
these constraints, we also need to introduce an intermediate generic
procedure instantiation so that we have:

1. The original generic procedure instantiation

2. A generic procedure that wraps 1.; this one can have a body, and thus
   host code to dump coverage buffers.

3. A generic procedure instantiation for 2., acting as the main for the
   instrumented project.

This commits implements this scheme to fix the handling of such mains.