Add the freestanding switch to the embedded build
The embedded build must have the minimal amount of dependency to system
headers, and may not even be executed in a context where a libc is
available. In this case, if the freestanding switch is specified, the
compiler will use its own implementation of stdint.h and will not assume
that one is provided by the system.