Remove the dependency to string.h in the core runtime
Bareboard runtimes may not provide the string.h header, but we can
reasonably assume that they will provide a definition for memcpy and
memset. Instead of relying on the string.h header, import their
definition manually. Do that in all of the files that belong to the
embedded configuration of the instrumentation runtime (what we call the
core). The full configuration comes with less build constraints, so no
need to do that there.

If these functions are not provided by the runtime, the user will have
to provide them: this is a pre-requisite of the instrumentation runtime.