Introduce exit/abort variants of the silent_last_chance_handler
The use of abort() to terminate was causing trouble in native
configurations as it was resulting in an error exit code from
tests raising unhandled exceptions on purpose.

We recently replaced it with calls to exit(0), relying on the
availability of a minimum libc in all BB configurations.

This caused other problems in BB environments using recent versions
of newlib which now invoque IO unconditionally from exit(), resulting
in undefined refs to write_r.

An incidental issue was the unconditional addition of an explicit
"-lc", which doesn't work on Windows.

This change mitigates the issue by improving testsuite framework
to feature two versions of the Silent_Last_Chance package body, one
using abort() and the other using exit(0), and select a variant
through a "naming" package in the generated project files.