Add sanity check for aspects attached to formal parameters
Aspects specified for formal subprogram parameters should be
instrumented in the same way as default expressions, so there is
nothing in particular to be done.

We also have no aspect that is supported by gnat and that accepts
an expression executed at runtime, so we don't have the possibility
to add a meaningful test for this.

This is a sanity check to ensure there are no crashes in gnatcov
while processing code with aspects attached to formal parameters.