Fix testsuite.py setting of TARGET variable for libsupport
The move to a more systematic build of libsupport brought
a change in the value used to set the TARGET variable, from
self.env.target.triplet to args.target just for consistency
with how other variables were set.

This was actually inappropriate as the variable is used as
a --target gprbuild switch and args.target may contain extensions
not fit for that use.

Revert to the previous source of value and to a more canonical
way to guard such references (self.env.is_cross).

Part of work for issue eng/cov/gnatcoverage#3