Reject files passed as SID that do not have a .sid extension
Previously all files given to --sid or listed in a file given to --sid
were accepted an treated as such. This meant that any error occuring
during the treatment of the file as a SID file was the only way to be
aware of a file being invalid. Now only files whose name ends with
".sid" are accepted as SID files and the error message is adaped to make
it explicit.

TN: U429-016