Improve messages upon misuse of --units
The check for the use of --units without -P could never be satified so
the appropriate error message was never printed, this is now fixed. A mention
of the -P options to use along with --units has been added to the error
message when there are no user provided SCOs.

TN: U429-016