instrument-input_traces.adb: fix handling of long lines in Base64 traces
Do not use the `Ada.Text_IO.Get_Line` function, as it returns a `String`
on the stack, so it may trigger stack overflows with too long lines. Use
instead the `Get_Line` procedure that uses a fixed-size buffer, and just
skip too long lines, which are never supposed to be part of source
traces anyway.