Genbundle: Perform consistency checks on status files
Genbundle used to perform consistency checks on the tc.dump
file, mostly containing the SCO counters for each testcase (as in
SCOV Testcase). There are tests in the Common directory that have
multiple SCOV Testcases per actual test (i.e. testcase as identified
in the TOR document), which made this consistency check impossible.

The consistency check is now done on the testcase status file,
tcs.dump, as only one is generated per test.