cutils.py: strip the tool version strings
When dumping the tool versions in the testsuite environment dump,
tool version strings are split by line on a "\n". This results,
when run on windows, in stray carriage returns being included in
the dump file, in turn messing up the formatting of the environment
tables in the STR document.

Fix this by consistently stripping each line of the version string.

Part of W426-013 (gnatcov qualkit 999-W426-013)