Identify units of file-based language by their fullname
We used to rely on the base name + project name, but incoming changes
wrt. Parallelized instrumentation will make it harder to identify the
project name. When instrumenting a single unit, we won't have any more
project information as we only load the project once: when instrumenting
the whole project tree, which then delegates the instrumentation of
units to subprocesses not loading the project again.

We could theoretically pass all this information on the command-line,
but it is not convenient and it goes against the end goal of integrated
instrumentation, aiming at supporting other build systems than gprbuild.

Also use this as an opportunity to tackle our handling of subunits. Add
them explicitly in the unit map.