Move code for compiler switches parsing to Instrument.Common
This is preparatory work for the parallelized instrumentation.

As compiler switches can be specified in a project file, and as we want
to drop the dependency to projects the most we can in the unit
instrumentation process, we will need to pass such switches before
instrumenting the unit. This means that functions used for compiler
switches parsing purposes must be exposed, and we would rather expose
them in the Instrument.Common unit, rather than in the Instrument.C unit
that is stubbed (and thus everytime we add public API, we have to stub it).