Fix Assertion_Policy in decision/Assertions/Inactive/TypeInvariants
The test verifies the tool behavior in a context
where a TypeInvariant is meant to be disabled, resulting
in an invariant check function never called.

The disabling is implicit in absence in -gnata (invariants
are not activated by default), and doubled by an explicit
Assertion_Policy pragma in case -gnata is in the compilation
flags.

The pragma needs to target the precise kind of aspect
attached to the declaration. We have

    type T_Value is tagged private
     with Type_Invariant'Class => I_Check(T_Value);

so the Assertion_Policy should target Type_Invariant'Class.

It was targetting "Invariant" so far and the test was
failing when compiled with -gnata.

Part of W426-013 (gnatcov qualkit 999-W426-013)