xnexpanders.py: Restrict the accepted delimiters for fuzzy blocks
Prior to this change, any tag on a line containing a lone
"begin" or "end" in a source of interest would be considered as a
start or end of a fuzzy block.

These are obsolete and not used anymore, but their processing is
quite deeply rooted is the testsuite sources, so simply restrict
the range of tags that can be used to define a fuzzy block.

This change also adapts one test that used to define a fuzzy block
by accident, and no longer works once the fuzzy block is removed.