Shorten filenames for generated instrumentation files
As we now add the hash of the fullname to have a unique slug for
compilation units identified by file fullnames (for file-based
languages, such as C / C++), generated instrumentation files (such as
coverage buffers) have longer names. This could be a problem on windows,
as the paths can't exceed a certain length.

To make the generated filenames shorter, we can both:

  * Use an hexadecimal image of the hash to have a shorter slug

  * Generate instrumentation files with a shorter prefix than the
  current one, which is "gnatcov_rts-buffers": we replace it with
  "gcvrt".