U204-026-arch-mix: kill test for light profiles
When using --dump-trigger=main-end on a native profile, we use
controlled objects and finalization to make sure that we do not miss any
code execution (notably the execution of finalizers for controlled
objects declared at the beginning of the main). This is not compatible
with cross targets under light profiles, as controlled objects are not
available in these configuration.

Kill the tests there, as its main goal is to check inter-host
compatibility. The limitation of having sources instrumented natively
not being compilable for cross light profiles is acceptable.