Improve the usability of manual dump trigger
When using the manual dump trigger option, one dump helper unit per
project is generated containing everything needed to perform a call to
its buffers dumping procedure anywhere in the code.  gnatcov will look
for the pragma statement 'pragma Annotate (Xcov, Dump_Buffers)' (for
Ada) or the comment '/* GNATCOV_DUMP_BUFFERS */ (for C) in all sources
and replace it with the actual call.