Add generation of coverage percentage indices
When generating xcov, xcov+ and report format reports coverage indices
are also generated, one for the enabled coverage level and one for the
lines metric. They contain the percentage of lines/obligations that fall
under each coverage state of each unit of interest, project and
aggregation of all projects.

Also, only log the creation of an output directory if quiet mode is
disabled or if the directory to be created is an output subdirectory.