Instrument Ada: Convert elsif expressions to Standard.Boolean
When instrumenting elsif expressions, we insert the statement witness
call as a function call in an OR ELSE expression with the original
expression. This may create a type mismatch as the Witness function
returns a Standard.Boolean, whereas the original expression may be
a derived type from Boolean.

This change fixes this by always converting the original expression
to Standard.Boolean.