Introduce subprograms of interest
With this change, users now have the ability for finer-grained filtering
- on a subprogram of interest basis. This changes actually provides more
than that: filtering per scopes, which are identified by the
<filename>:<line> of their specification declaration point for Ada, if
they have any (otherwise it is the body declaration point), and by the
body declaration point for C / C++.

The subprograms of interest filtering is applied after all of the units
of interest filtering has been applied. This means that if the users
passes --subprograms=<ignored_file>:<line>, it won't be in the coverage
report.

To implement this, we use the logic implemented to support scope
metrics, and refactor it to use trees instead of ad-hoc data structures.