Ada instrumenter: decide to skip instr. a decision in Process_Decisions
Before this patch, the decision of whether to instrument an identifier
decision is taken after the low-level to high-level SCOs conversion in
`Instrument_Source_File`. Upcoming work will require this decision to be
aware of the expression context (in entry guard?), and it is more
natural to express this in a place where the context is available: in
`Process_Decisions`.

This is just a refactoring: no change of behavior expected.