Remove the --analyze-entry-guards option
Never ignore 'G' decisions in the low-level to high-level SCO
conversion, and in the instrumenter, detect restrictions in order to
decide whether entry guards should be instrumented as decisions. When
they are not instrumented, flag these decisions as non-instrumented
instead of just ignoring them.

As a side effect, this fixes the instrumentation of decisions for select
alternative barriers (`select ... when X =>`), which used to crash the
Ada instrumenter.

On configurations that support non-pure/simple barriers, this creates
new coverage obligations for entry barriers in user code: adjust
existing tests accordingly.