Add test for container aggregates and generalized array aggregates
The test for container aggregates does nothing useful, in a convoluted
way simply for the sake of introducing decisions in the expressions of
the keys of the aggregates. The test also fails as Libadalang doesn't
support the `[use Key_Expression]` syntax as of yet, so gnatcov fails to
instrument the source file.