C instr: fix postprocessing wrt. pragma directives
As a reminder, we postprocess the preprocessed code to avoid nesting in
the SCOs source location (as we rely on the preprocessor line
directives), e.g. when we have occurrences of the following:

```
# 1 main.c
int a = foo();  // SCO #1 span is [1:14]
# 1 main.c
bool b = bar(); // SCO #2 span is [1:15], CONFLICT
```

we rewrite them to

```
# 1 main.c
int a = foo(); bool b = bar();
// SCO #1 span is [1:14] and SCO #2 is [16:30], no conflict
```

This is proven to be an issue when there are pragma directives, e.g.

```
# 1 main.c
#pragma GCC diagnostic push
# 1 main.c
#pragma GCC diagnostic ignored "-Wnon-null"
```

as we rewrite them to

```
# 1 main.c
#pragma GCC diagnostic push#pragma GCC diagnostic ignored "-Wnon-null"
```

which is invalid code.

We thus make an exception for preprocessing directives as they can't
yield coverage obligations, and thus won't trigger the issue mentionned
above.