C/C++ instrumentation: include externally built project source dirs
To preprocess a C/C++ source file, we need to locate the source
directories that will be ultimately pass down through -I switches to the
compiler driver command line.

In a GPR context, a source can reference a header in any of the project
dependencies source directories. We thus do a full traversal of the
project tree and register every source directory. Nevertheless, we were
not traversing externally built projects, as it depended on the setting
of the Externally_Built_Projects_Processing_Enabled variable. Refactor
the code to be able to traverse through externally built project in that
case.