C/C++ instr: consider other compiler switches for preprocessing
We must launch the preprocessing command with all of the switches that
can influence the preprocessing of the file. We used to consider macro
definition / undefinition and the C/C++ standard set by the -std switch
but there are more switches than that (the list is unfortunately
probably not exhaustive): -fno-rtti (do not include runtime type
information) and all of the warnings switches which can influence the
preprocessing through the __has_warning macro.