Instrument.Common: Do not warn for big BDD in stmt+decision
gnatcov emits a warning when encountering a decision with too many
paths to keep track of. This warning is not relevant for decision
coverage as the non instrumentation of the individual conditions
does not affect the accuracy of the coverage report for level
stmt+decision.

This change removes the warning when MC/DC coverage is not enabled.