Instrument.Ada_Unit: refactor instrumentation process
We no longer process statement by sequences and defer the processing of
decisions. This was a leftover from the computation of dominance
information, which is not needed in instrumentation mode, and it
simplifies the instrumentation process a bit.

We also use this as an opportunity to refactor the way we deal with non
instrumented source code constructs, and make it similar to what is done
in the C instrumentation process.