Instrument.C: fix instrumentation when no SCO can be instrumented
gnatcov used to crash when instrumenting source code that consisted
solely of non-instrumentable constructs. This was due to an implicit
assumption that every source of interest should be present in the
UIC.Instrumented_Entities map, which wasn't the case in this scenario.
Fix the code to ensure the assumption's validity.