clang-wrapper.cc: fix clang_getCalleeName
The callee declaration can be, e.g. a parameter declaration if the call
expression is a parameter callback, which is not a derived type of
FunctionDecl, and thus can't be casted to it. Only deal with
FunctionDecl (and FunctionTemplateDecl) type derivations and return an
empty strings for the other cases.