doc: remove outdated references to `gnatcov_rts_full`
We no longer have two versions of the runtime, but one version with
various build configuration, and a `gnatcov setup` command picking the
right configuration according to the target switches.

Also document the build + install of the coverage runtime using `gnatcov
setup` (which is the standard process) rather than manually building and
installing the runtime.