Fix line endings for instrumented sources under windows
We need to open and write the instrumented source as a binary file, to
be consistent with Libadalang which retrieves the tokens from a file
opened in binary mode. It provides us with line terminators when
unparsing the user-defined tokens and we must be careful to output these
unaltered. On Windows for example, we get CRLF line terminators in the
LAL tokens, as CR is not automatically stripped when opening a file in
binary mode. If these line terminators (CRLF) were written in a text
file, calls to fwrite would replace LF with CRLF, resulting in CRCRLF,
which is incorrect.