Fix handling of lambda expressions
When retrieving the lambda expressions inside a clang cursor, we were
traversing recursively all the nodes in the cursor subtree. In addition
to being very inefficient, this resulted in lambda expressions being
processed multiple times.

To fix this, we integrate the processing of lambda expression to the
processing of decisions, similarly to what is done to process declare
expressions in Ada.