last_chance_handler: fix insertion of with clause
To insert the 'with' clause for the last chance dump unit, we picked the
first line that was not a pragma. This relied on the implicit assumption
that we did not have code where a line was composed both of a pragma and
the start of a package / procedure, e.g.

pragma Style_Checks (Off); procedure Foo is ...

As we can now have such code occurrences in instrumented sources,
because we no longer introduce new lines for inserted Style_Checks and
Warnings pragmas, we drop this heuristic and use LAL syntax tree to know
where to insert the with clause.