Instrument.C: fix call to Import_Options
The Analysis_Options record stores the list of options useful for
preprocessing purposes, among them being a mapping of units to compiler
switches (that can alter the preprocessing). This mapping maps the
original source filename to its list of compiler switches, and not the
preprocessed filename. Fix it to pass the former instead of the latter.