Instrument.C: fix manual dump-trigger
we used to have mismatching coverage obligations between the step
recording preprocessing information and the step instrumenting because
we were missing standard preprocessor search paths when parsing the
source file in the former step.

For recall, to record preprocessing information, we parse the
unpreprocessed file with clang, and we inhibit builtin macros, as well
as standard compiler headers to fully mimick the behavior of the
preprocessor _in use_ (e.g. gcc). This means that if we are missing part
of that configuration, which was the case there, we end up with an
ill-formed AST. This results in wrong coverage obligations for the step
recording preprocessing information.