Instrument Ada: add support for preprocessing
Extract preprocessor configuration data from project files and pass it
to Libadalang, so that instrumentation works on preprocessed code.

Naturally, this allows gnatcov to compute the code coverage only for
code that is left "enabled" by preprocessing directives: disabled code
(i.e. equivalent to `#if false`) is ignored and thus creates no coverage
obligation. Note that consolidation will not help including code from
all "preprocessing branches" in coverage reports, as gnatcov requires
(and checks) that coverage obligations are the same for two units to
consolidate.

Similarly, the coverage obligations for code that comes from symbol
expansion (`$Foo` expanded into `Bar` with `-Dfoo=Bar`) designates
expanded code. Even though line numbers are preserved during
preprocessing, column numbers may be different between original code and
preprocessed code and thus coverage reports. This is a limitation from
Libadalang's preprocessing facilities, so not something we could fix at
the gnatcov level, but this limitation is not considered problematic at
this point for the known use cases of coverage on code with
preprocessing directives.