Instrument.adb: Disambiguate hash-based slugs for Ada units
Change the string from which the hash is computed when generating
a slug for an Ada unit to avoid collisions.

The hash was computed from the symbol string of the unit name, but
this transformation can result in the same symbol name being produced
for two units: Foo.Bar and Foo_Bar will both be turned into foo_bar.

Instead, use the human readable slug as a basis to compute the hash.