SC_Obligations: assert SCOs nesting/ordering invariants for scopes
The scope entities traversal code assumes that SCOs ranges in scope
entities (Scope_Entity.From/.To) are correctly nested/ordered in scope
entities trees. Add code to check that this invariant is respected each
time we create/modify these trees.

These invariants are supposed to hold no matter how gnatcov is used and
checking them is not trivial, so check them only when assertions are
enabled.