SC_Obligations: rework the scopes querying API
Exactly one set of subprograms of interest is ever used: the set for
--subprograms command line arguments. Remove Is_Active's corresponding
argument and always use the set at Switches.Subps_Of_Interest instead.
This also opens the way to more efficient ways to check whether the
current SCO is covered by a scope that matches --subprograms.

Also rename Is_Active to avoid the confusion with the
Active_Scopes/Active_Scope_Ent components (which can contain scopes that
are *not* active according to --subprograms arguments).