instrument-c.adb: avoid uninitialized record component
Make sure that the record component PP_Info.PP_Source_Range is always
initialized, at least so that we do not stream absurd sloc range
information to SID/checkpoint files.

That went unnoticed so far because Natural'Write will happily write a
negative number, and Natural'Read will happily read and assign it,
without raising a Constraint_Error.

This fixes no known gnatcov issue as of today, but is necessary for the
upcoming work to make SID/checkpoint reading manual.