libsupport.gpr: minimize C language inclusion
In preparation to the introduction of a CCG testsuite run, minimize
the cases in which C is listed as a required language in
libsupport.gpr.

CCG does not support compiling projects with C sources. As such, only
include the C language in a cross targeted scenario, when there actually
are C files to be compiled.