Makefile.Libsupport: add more customization options
This change allows the customization of the exit mechanism in the last
chance handler, and to pass compilation flags when building the support
library.

This is in preparation of the introduction of a CCG testsuite run.
The heuristics to choose which compilation flags need to be passed to
gprbuild, and which exit mechanism needs to be used in the last chance handler
do not work in a CCG context. Instead of adding corner cases for each
configuration, allow configuring these features during the make invocation.