Testsuite: Allow altruns to specify an alternate builder
The altrun mechanism allows to specify alternate programs for all the
"gnatcov <command>" commands. To be able to support CCG, we also need
to be able to override which builder is used.

This change adds the possibility to specify an overriding program for
gprbuild.

Note that gprbuild still needs to be present on path, as it will be used
when dumping version information, and to build the support library, as
this is done through a Makefile out of the testsuite's control.