Testsuite: Add an altrun for CCG
This altrun allows the testsuite to build test executables through
CCG, for a native target.

The altrun inlcudes:
 - a pre_testsuite.py script that prepares the coverage runtime,
   the support library and the Ada runtime;
 - a cgprbuild.py script that is meant to replace gprbuild. It invokes
   CCG through gprbuild, then collects all the generated C files and
   compiles them to generate an executable;
 - various other files used by the two above scripts as build helpers.

Note that CCG does not support compiling projects that have C listed as
a project language, the binder will thus not execute and no main will be
generated. As such, it is best not to run the full testsuite with this
altrun but only a subset of tests which only contains Ada sources, such
as Qualif/Ada.