gnatcov_rts.gpr: Add configuration for use with CCG toolchain
The CCG toolchain does not support compiling C code. As such,
in order to be able to use GNATcov_RTS with this toolchain, there
must be no C sources in the project, and no references to C files.

This change introduces a new configuration variable, For_CCG, which
when active removes C from the source languages, and removes C source
exclusions.