Ada2022/aggregates/containers: workaround the compiler hang
When compiling the instrumented code for this testcase, GNAT is
currently stuck in an infinite loop. This is highly inconvenient for the
GNATcov team, as this means that this testcase always hits the timeout,
making affected testuite runs take ~13 minutes more than needed.

Temporarily make the testcase fail instantly while the GNAT bug is being
investigated.