SCOV.map: do not pass "-v" to "gnatcov map-routines"
The goal of this test driver is to detect the presence of decision map
issue ("!!!" and "***" messages), which are emitted regardless of
whether "-v" is passed.

Upcoming work will trigger the emission of "branch stats" when "-v" is
passed (it is currently enabled only when --branch-stats is passed),
and logging code for --branch-stats currently crashes or create spurious
test failures. As a workaround, do not pass "-v" anymore.