manual dump: fix performance issue
When looking for dump trigger indications in a C/C++ source, we were so
far reading the source file contents to an unbounded string, which is
very inefficient for large files. Use the GNATCOLL.Mmap API instead to
avoid the costly string copies/allocations.

This was noticed when instrumenting gnatcov with the manual
dump-trigger, and notably when trying to look for a manual dump
indication in the `clang-wrapper.cc` file. The preprocessed version of a
C++ file including parts of the standard library can be quite large,
which incurred this performance slow down.