C/C++ instr: use the raw list of switches to preprocess the sources
We currently have a selective list of switches that are passed on to the
preprocessing invocation, which has proven to not be a scalable
solution (the number of switches that can influence the preprocessing,
notably the source search paths is pretty huge).

This patch opts for a slightly different approach. As a reminder, we
need to preprocess both with (1) the original compiler driver, and (2)
with clang.

We will now pass on all of the switches when preprocessing with the
original compiler driver, without applying any filtering. That way we
are ensured that this preprocessing will be correct, and will not be
missing part of the user configuration, which could be the case before.

We will still keep the selective list of switches to have control over
the switches that are passed on to the clang preprocessing invocation.
We must do that as, e.g. some gcc switches are not accepted by clang.
Note that we do not _only_ use user specified switches to, e.g. know the
source search paths: we also use the preprocessing output of (1).

We retrieve, e.g. the list of search paths from this preprocessing
output, so we expect that in practice, improving (1) will also make (2)
more robust. And in the worst case scenario, the preprocessing step will
yield different results for (1) and (2), in which case we will simply
discard preprocessing information, which is what (2) is for.